+++ CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 ----
 
  * `Indexer\SourceIndexer` now passes its options to the loader as well as to the document processor
+ * [BC BREAK] Accept and return `Platform\Vector\VectorInterface` in `Query\VectorQuery` and `Query\HybridQuery` instead of the final `Platform\Vector\Vector` class
+
 
 0.11
 ----